LUN-3: Move Interfaces from profession.ts
