--- src/interfaces/profession.ts
@@ -1,11 +1,0 @@
-export interface IProfessionScreenProps {
-  navigation: any
-}
-
-export interface IProfessionsProps {
-  id: number
-  title: string
-  description: string
-  icon?: string
-  total_training_sets: number
-}
